refactor: Revert entity IDs to Integer and fix ProjectAddress CRUD
This commit reverts all entity ID fields from `Long` back to `Integer` as requested. This change has been applied to all affected models, repositories, DTOs, services, controllers, and tests.

It also fixes the implementation of the `ProjectAddress` CRUD to support a one-to-many relationship, as pointed out in the code review.

Key changes:
- Reverted all entity IDs from `Long` to `Integer`.
- Updated all related tests to use `Integer` for IDs.
- Updated the `Project` entity to have a one-to-many relationship with the `Address` entity.
- Updated the `ProjectAddressService` and `ProjectAddressController` to handle multiple addresses for a project.

+++ src/main/java/com/management/person/service/PersonService.java
@@ -108,13 +108,13 @@ public class PersonService {
                 .collect(Collectors.toList());
     }
 
-    public PersonDTO findById(Long id) {
+    public PersonDTO findById(Integer id) {
         Person person = personRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Pessoa não encontrada com o id: " + id));
         return modelMapper.map(person, PersonDTO.class);
     }
 
-    public List<PersonDTO> findPersonsByCompanyId(Long companyId) {
+    public List<PersonDTO> findPersonsByCompanyId(Integer companyId) {
         return companyPartnerRepository.findByCompanyId(companyId)
                 .stream()
                 .map(companyPartner -> modelMapper.map(companyPartner.getPerson(), PersonDTO.class))
