+++ src/main/java/com/management/resource/controller/ProjectResourceController.java
@@ -4,13 +4,13 @@ import com.management.resource.dto.ProjectResourceCreateRequest;
 import com.management.resource.dto.ProjectResourceUpdateRequest;
 import com.management.resource.dto.ProjectResourceResponse;
 import com.management.resource.service.ProjectResourceService;
+import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import io.swagger.v3.oas.annotations.Operation;
 
 import java.util.List;
 
@@ -30,25 +30,25 @@ public class ProjectResourceController {
 
     @GetMapping("/resources/{id}")
     @Operation(summary = "Finds a resource by ID")
-    public ResponseEntity<ProjectResourceResponse> getById(@PathVariable Long id) {
+    public ResponseEntity<ProjectResourceResponse> getById(@PathVariable Integer id) {
         return ResponseEntity.ok(resourceService.findById(id));
     }
 
     @GetMapping("/projects/{projectId}/resources")
     @Operation(summary = "Finds all resources for a project")
-    public ResponseEntity<List<ProjectResourceResponse>> getByProject(@PathVariable Long projectId) {
+    public ResponseEntity<List<ProjectResourceResponse>> getByProject(@PathVariable Integer projectId) {
         return ResponseEntity.ok(resourceService.listByProject(projectId));
     }
 
     @PutMapping("/resources/{id}")
     @Operation(summary = "Updates a resource")
-    public ResponseEntity<ProjectResourceResponse> update(@PathVariable Long id, @Valid @RequestBody ProjectResourceUpdateRequest request) {
+    public ResponseEntity<ProjectResourceResponse> update(@PathVariable Integer id, @Valid @RequestBody ProjectResourceUpdateRequest request) {
         return ResponseEntity.ok(resourceService.update(id, request));
     }
 
     @DeleteMapping("/resources/{id}")
     @Operation(summary = "Deletes a resource")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         resourceService.delete(id);
         return ResponseEntity.noContent().build();
     }
